requirements.txt
django<2
dj-database-url
conversion
django-requestlogging
django-log-request-id
psycopg2
waitress
django-extensions
djangorestframework
ipython<6
django-debug-toolbar==1.8
django-mass-edit
django-import-export
django-nested-admin
python-google-places
django-sendgrid-v5
enum34
celery<5
google-python-cloud-debugger
google-cloud-bigquery
django-model-utils
django-audit-trail
django-auto-one-to-one
python-dateutil
redis
django-jet
django-favicon
django-heroku-memcacheify
whitenoise>3.3.0
gunicorn
django-redis
raven
newrelic
django-admin-easy==0.4.1
django-admin-view-permission
django-admin-rangefilter
django-inline-actions
